Make 'get' and 'delete' use branch protection rules
Change 'get' and 'delete' implementation deal with
branch protection rules instead of branch protections.
The commands now accept a rule ID from the user.
The ID can be found with the 'list' command.
This simplify a lot the code and is more aligned with
the GitHub API abstraction.

This also fix an open bug/limitation as now the extension
supports the situation where you have a rule but no existing branch
matches the rule.